+++ src/renderer/helpers/index.ts
@@ -7,3 +7,4 @@ export * from './midi';
 export * from './note';
 export * from './chords';
 export * from './suggestions';
+export * from './autoContext';